# api/swagger/swagger.yaml
basePath: /api/v1
definitions:
  service.CreateMessageRequest:
    properties:
      payload: {}
      tenant_id:
        type: string
    type: object
  service.CreateTenantRequest:
    properties:
      name:
        type: string
      workers:
        type: integer
    type: object
  service.CursorPaginationResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/service.MessageResponse'
        type: array
      has_more:
        type: boolean
      next_cursor:
        type: string
    type: object
  service.MessageResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      payload: {}
      retry_count:
        type: integer
      status:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
  service.TenantResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      workers:
        type: integer
    type: object
  service.UpdateConcurrencyRequest:
    properties:
      workers:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: A multi-tenant messaging system with RabbitMQ and PostgreSQL
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Multi-Tenant Messaging System API
  version: "1.0"
paths:
  /messages:
    get:
      description: Get messages with cursor pagination
      parameters:
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      - description: Limit (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Filter by tenant
        in: query
        name: tenant_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.CursorPaginationResponse'
      summary: List messages
      tags:
      - messages
    post:
      consumes:
      - application/json
      description: Publish a message to a tenant's queue
      parameters:
      - description: Message creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/service.MessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Publish a message
      tags:
      - messages
  /messages/{id}:
    get:
      description: Get details of a specific message
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: query
        name: tenant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.MessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get message details
      tags:
      - messages
  /tenants:
    get:
      description: Get a list of all tenants
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.TenantResponse'
            type: array
      summary: List all tenants
      tags:
      - tenants
    post:
      consumes:
      - application/json
      description: Create a new tenant and spawn a message consumer
      parameters:
      - description: Tenant creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/service.TenantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new tenant
      tags:
      - tenants
  /tenants/{id}:
    delete:
      description: Delete a tenant and stop its consumer
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a tenant
      tags:
      - tenants
    get:
      description: Get details of a specific tenant
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TenantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get tenant details
      tags:
      - tenants
  /tenants/{id}/config/concurrency:
    put:
      consumes:
      - application/json
      description: Update the number of workers for a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Concurrency update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateConcurrencyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TenantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update tenant concurrency
      tags:
      - tenants
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"